chore: Refactored repository layer, to have a single source of truth, provide data from local data source
