+++ app/src/main/java/com/mp/cleanmovie/app/navigation/AppNavHost.kt
@@ -7,7 +7,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.mp.cleanmovie.movieDetails.presentation.MovieDetailScreen
+import com.mp.cleanmovie.movieDetail.presentation.MovieDetailScreen
 import com.mp.cleanmovie.movieList.presentation.MovieListScreen
 
 @Composable
