+++ app/src/main/java/com/mp/cleanmovie/core/di/LocalModule.kt
@@ -1,4 +1,5 @@
 package com.mp.cleanmovie.core.di
+
 import android.content.Context
 import androidx.room.Room
 import com.mp.cleanmovie.movieList.data.local.database.MovieDao
@@ -12,15 +13,15 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
- object LocalModule {
-        @Provides
-        @Singleton
-        fun provideDatabase(
-            @ApplicationContext context: Context
-        ): MovieDataBase {
-            return Room.databaseBuilder(context, MovieDataBase::class.java, "movie_database.db").build()
-
+object LocalModule {
+    @Provides
+    @Singleton
+    fun provideDatabase(
+        @ApplicationContext context: Context,
+    ): MovieDataBase {
+        return Room.databaseBuilder(context, MovieDataBase::class.java, "movie_database.db").build()
     }
+
     @Provides
     fun provideMovieDao(movieDataBase: MovieDataBase): MovieDao {
         return movieDataBase.dao
