+++ app/src/main/java/com/mp/cleanmovie/movieDetail/di/MovieDetailBinderModule.kt
@@ -1,9 +1,9 @@
-package com.mp.cleanmovie.movieDetails.di
+package com.mp.cleanmovie.movieDetail.di
 
-import com.mp.cleanmovie.movieDetails.data.local.datasource.MovieDetailLocalDataSource
-import com.mp.cleanmovie.movieDetails.data.local.datasource.MovieDetailLocalDataSourceImpl
-import com.mp.cleanmovie.movieDetails.data.repository.MovieDetailRepositoryImpl
-import com.mp.cleanmovie.movieDetails.domain.repository.MovieDetailRepository
+import com.mp.cleanmovie.movieDetail.data.local.datasource.MovieDetailLocalDataSource
+import com.mp.cleanmovie.movieDetail.data.local.datasource.MovieDetailLocalDataSourceImpl
+import com.mp.cleanmovie.movieDetail.data.repository.MovieDetailRepositoryImpl
+import com.mp.cleanmovie.movieDetail.domain.repository.MovieDetailRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
