+++ app/src/main/java/com/mp/cleanmovie/movieDetail/domain/repository/MovieDetailRepository.kt
@@ -1,4 +1,4 @@
-package com.mp.cleanmovie.movieDetails.domain.repository
+package com.mp.cleanmovie.movieDetail.domain.repository
 
 import com.mp.cleanmovie.core.domain.model.DomainMovieData
