+++ app/src/main/java/com/mp/cleanmovie/movieList/data/repository/MovieListRepositoryImpl.kt
@@ -1,35 +1,39 @@
 package com.mp.cleanmovie.movieList.data.repository
 
-import com.mp.cleanmovie.movieList.data.remote.model.MovieData
-import com.mp.cleanmovie.movieList.data.local.database.MovieEntity
-import com.mp.cleanmovie.movieList.data.local.MovieListLocalDataSource
-import com.mp.cleanmovie.movieList.data.remote.MovieListRemoteDataSource
 import com.mp.cleanmovie.core.domain.model.DomainMovieData
+import com.mp.cleanmovie.movieList.data.local.MovieListLocalDataSource
+import com.mp.cleanmovie.movieList.data.local.database.MovieEntity
+import com.mp.cleanmovie.movieList.data.remote.MovieListRemoteDataSource
+import com.mp.cleanmovie.movieList.data.remote.model.MovieData
 import com.mp.cleanmovie.movieList.domain.repository.MovieListRepository
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.map
-
 import javax.inject.Inject
 
-class MovieListRepositoryImpl @Inject constructor(
-    private val movieListRemoteDataSource: MovieListRemoteDataSource,
-    private val movieListLocalDataSource: MovieListLocalDataSource,
-    private val mapper: Mapper
+private const val TAG = "MovieListRepositoryImpl"
 
+class MovieListRepositoryImpl @Inject constructor(
+    private val remoteDataSource: MovieListRemoteDataSource,
+    private val localDataSource: MovieListLocalDataSource,
+    private val mapper: Mapper,
 ) : MovieListRepository {
     override suspend fun getMovies(): Flow<List<DomainMovieData>> {
-        movieListRemoteDataSource.getMovies().collectLatest { remoteData ->
-            val mappedEntities = mapper.mapMovieDataListToMovieEntityList(remoteData.movies)
-            movieListLocalDataSource.insertMovies(mappedEntities)
-        }
-        return movieListLocalDataSource.getMovies().map {
-            mapper.mapEntityListToDomainList(it)
-        }
+        return localDataSource
+            .getMovies()
+            .map {
+                if (it.isEmpty()) {
+                    remoteDataSource.getMovies().collect { remoteData ->
+                        val mappedEntities =
+                            mapper.mapMovieDataListToMovieEntityList(remoteData.movies)
+                        localDataSource.insertMovies(mappedEntities)
+                    }
+                }
+                it.map { mapper.mapEntityToDomain(it) }
+            }
     }
 
     override suspend fun getSelectedMovie(id: String): DomainMovieData {
-        val movieEntity = movieListLocalDataSource.getSelectedMovie(id)
+        val movieEntity = localDataSource.getSelectedMovie(id)
         return mapper.mapEntityToDomain(movieEntity)
     }
 }
@@ -51,12 +55,11 @@ class Mapper @Inject constructor() {
                     released_on = released_on!!,
                     slug = slug!!,
                     genres = genres!!,
-                    cast = cast!!
+                    cast = cast!!,
                 )
             }
         }
     }
-
 
     fun mapEntityListToDomainList(movieEntity: List<MovieEntity>): List<DomainMovieData> {
         return movieEntity.map {
@@ -74,7 +77,7 @@ class Mapper @Inject constructor() {
                     released_on = released_on,
                     slug = slug,
                     genres = genres,
-                    cast = cast
+                    cast = cast,
                 )
             }
         }
@@ -95,7 +98,7 @@ class Mapper @Inject constructor() {
                 released_on = released_on,
                 slug = slug,
                 genres = genres,
-                cast = cast
+                cast = cast,
             )
         }
     }
